Update
Update consts for more accurate and clear API calls.
Separate the server socket from the activity message, call the server socket to listen to the message as soon as it enters the main interface. Reconnect when the connection to the socket server is lost.
Call API like, dislike, agree to friend, disagree to friend, message with friends.